+++ src/model/index.ts
@@ -1,0 +1,5 @@
+export * from "./natural-number";
+export * from "./non-empty-string";
+export * from "./pokemon";
+export * from "./pokemon-type";
+export * from "./whole-number";